Submit get request to scrape items
