+++ app/controllers/scraper_results_controller.rb
@@ -5,7 +5,8 @@ class ScraperResultsController < ApplicationController
     results = scraper.new(query: query).scrape
     results = FuzzySubstringMatchFilter.new(results, query).matches
     results = results.map { |r| OpenStruct.new(r) }
-    render partial: 'scraper_results', locals: { results: results }, layout: false
+    render json: results.to_json
+    # render partial: 'scraper_results', locals: { results: results }, layout: false
   end
 
   def import
