+++ app/javascript/bundles/Lists/components/Spinner.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const Spinner = () => (
+  <div className="columns">
+    <div className="column is-1 is-offset-5">
+      <span className="icon is-large">
+        <i className="fa fa-spinner fa-pulse fa-3x"></i>
+      </span>
+    </div>
+  </div>
+)
+
+export default Spinner
